% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_designs.R
\name{compare_designs}
\alias{compare_designs}
\title{compare_designs}
\usage{
compare_designs(
  EffSize,
  sigma,
  tau,
  family = gaussian(),
  timepoints = NULL,
  N = NULL,
  sig.level = 0.05,
  DesMat = NULL,
  Cl = NULL,
  trtDelay = NULL
)
}
\description{
a short wrapper that compares parallel, sw, and parallel+ baseline design
}
