% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_designs.R
\name{compare_designs2}
\alias{compare_designs2}
\title{Title}
\usage{
compare_designs2(
  Designs,
  mu0,
  mu1,
  family = gaussian(),
  sigma,
  tau,
  eta = NULL,
  rho = NULL,
  sig.level = 0.05
)
}
\arguments{
\item{Designs}{different designs to compare}
}
\value{

}
\description{
Title
}
