% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InformationContent.R
\name{compute_InfoContent}
\alias{compute_InfoContent}
\title{Title  Formula-based calculation of information content}
\usage{
compute_InfoContent(CovMat = NULL, W = NULL, dsn, sumCl, tp)
}
\arguments{
\item{CovMat}{#' @param CovMat numeric, a positive-semidefinite matrix with
(#Clusters \eqn{\cdot} timepoints) rows and columns.}

\item{W}{numeric, the inverse of a covariance matrix. If CovMat is specified,
input for W is ignored}

\item{dsn}{a matrix with  (#Clusters \eqn{\cdot} #timepoints) rows and p
columns, where p are the degrees of freedom of fixed effects in a gls model.
This usually contains the intervention effect and some specification of the
time effect.}

\item{sumCl}{number of clusters}

\item{tp}{number of time points}
}
\value{
A matrix containing the information content for every cluster-period cell
}
\description{
Title  Formula-based calculation of information content
}
