% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_CovMat.R
\name{construct_CorBlk}
\alias{construct_CorBlk}
\title{'construct_CorBlk'}
\usage{
construct_CorBlk(timepoints, sigma, tau)
}
\arguments{
\item{timepoints}{not implemented}

\item{sigma}{numeric, residual error of cluster means if no N given.
Else residual error on individual level}

\item{tau}{numeric, standard deviation of random intercepts}
}
\description{
'construct_CorBlk'
}
