% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_CovMat.R
\name{construct_CorMat}
\alias{construct_CorMat}
\title{'construct_CorBlk'}
\usage{
construct_CorMat(I, timepoints = NULL, sigma, tau, family = gaussian(),
  N = NULL)
}
\arguments{
\item{I}{integer (vector), number of clusters per wave (in SWD)}

\item{timepoints}{not implemented}

\item{sigma}{numeric, residual error of cluster means if no N given.
Else residual error on individual level}

\item{tau}{numeric, standard deviation of random intercepts}

\item{family}{distribution , not implemented}

\item{N}{integer (vector), number of individuals per cluster.
Defaults to 'rep(1,sum(I))' if not passed.}
}
\description{
'construct_CorBlk'
}
