% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_DesMat.R
\name{construct_trtvec}
\alias{construct_trtvec}
\title{construct_trtvec}
\usage{
construct_trtvec(Cl, trt_delay, design, timepoints)
}
\arguments{
\item{Cl}{integer (vector), number of clusters per wave (in SWD)}

\item{trt_delay}{numeric (possibly vector), value(s) between 0 and 1 specifing the
intervention effect in the first (second ... ) intervention phase}

\item{design}{character, specifies the study design. Defaults to "SWD".}

\item{timepoints}{numeric, scalar}
}
\value{
a vector trtvec
}
\description{
construct_trtvec
}
\examples{


}
