% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunction.R
\docType{data}
\name{EffSize}
\alias{EffSize}
\title{'wlsMixedPower'}
\format{An object of class \code{numeric} of length 1.}
\usage{
EffSize
}
\arguments{
\item{EffSize}{numeric, raw effect}

\item{I}{integer (vector), number of clusters per wave (in SWD)}

\item{sigma}{numeric, residual error of cluster means if no N given.
Else residual error on individual level}

\item{tau}{numeric, standard deviation of random intercepts}

\item{sig.level}{numeric, significance level, defaults to 0.05}

\item{delay}{numeric (possibly vector), value between 0 and 1 specifing the
intervention effect in the first (second ... ) intervention phase  *not implemented*}
}
\description{
'wlsMixedPower'
}
\keyword{datasets}
