% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_alloc_Power.R
\name{plot.N_alloc_Power}
\alias{plot.N_alloc_Power}
\title{plot.N_alloc_Power}
\usage{
\method{plot}{N_alloc_Power}(x, bins = 30, show_mean = TRUE, show_IQR = TRUE, fill = "steelblue", ...)
}
\arguments{
\item{x}{An object of class \code{N_alloc_Power}}

\item{bins}{integer, number of histogram bins, spanning the observed
power range plus a small padding.}

\item{show_mean}{logical, should a dashed vertical line at the mean power
be added?}

\item{show_IQR}{logical, should a shaded band over the interquartile range
(Q1 to Q3) be added behind the bars?}

\item{fill}{character, colour of the histogram bars. Defaults to
"steelblue", matching the package's colour scheme.}

\item{...}{Arguments to be passed to methods}
}
\value{
a plotly html widget, displaying the power distribution as a
histogram with reference lines for the mean and the interquartile range.
}
\description{
plot.N_alloc_Power
}
\examples{
ap <- N_alloc_Power(N_pool = c(12, 8, 10, 9, 14),
                    Cl = rep(1, 5),
                    mu0 = 0, mu1 = 1, sigma = 2, tau = 0.33)
plot(ap)

}
