% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlsPower.R
\name{plot.wlsPower}
\alias{plot.wlsPower}
\title{plot an object of class `wlsPower`}
\usage{
\method{plot}{wlsPower}(x, ...)
}
\arguments{
\item{x}{object of class wlsPower}

\item{...}{Arguments to be passed to methods}
}
\value{
a plotly html widget
}
\description{
Plot a matrix that visualises the influence of each cluster for
each timepoint
}
