% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlsPower.R
\name{plot_CellWeights}
\alias{plot_CellWeights}
\title{plot cell contributions (weights) of a wls object}
\usage{
plot_CellWeights(
  x,
  annotations = NULL,
  show_colorbar = TRUE,
  marginal_plots = TRUE
)
}
\arguments{
\item{x}{object of class wlsPower}

\item{annotations}{logical, should the cell contributions be annotated in the Plot?}

\item{show_colorbar}{logical, should the colorbars be shown?}

\item{marginal_plots}{should the influence of whole periods, clusters also be plotted?}
}
\value{
a plotly html widget
}
\description{
plot cell contributions (weights) of a wls object
}
