% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_CovMat.R
\name{plot_CovMat}
\alias{plot_CovMat}
\title{Visualise a Covariance Matrix}
\usage{
plot_CovMat(CovMat, show_colorbar = TRUE)
}
\arguments{
\item{CovMat}{A covariance matrix (possibly in sparse matrix notation)}

\item{show_colorbar}{logical, should the colorbar be shown?}
}
\value{
a plotly object
}
\description{
Currently not exported.
}
