% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlsPower.R
\name{plot_InfoContent}
\alias{plot_InfoContent}
\title{plot the information content of a wls object}
\usage{
plot_InfoContent(
  IC,
  annotations = NULL,
  show_colorbar = TRUE,
  marginal_plots = TRUE
)
}
\arguments{
\item{IC}{a matrix with information content for each cluster at each time period}

\item{annotations}{logical, should the cell contributions be annotated in the Plot?}

\item{show_colorbar}{logical, should the colorbars be shown?}

\item{marginal_plots}{should the influence of whole periods, clusters also be plotted?}
}
\value{
a plotly object
}
\description{
plot the information content of a wls object
}
