% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wlsPower.R
\name{plot_wlsPower}
\alias{plot_wlsPower}
\title{'plot_wlsPower'}
\usage{
plot_wlsPower(wlsPower)
}
\arguments{
\item{wlsPower}{Output of function wlsMixedPower}
}
\value{
a formula for CoxEval()
}
\description{
'plot_wlsPower'
}
\examples{
BuildFormula(VarName="Variable",AdjName="additiveAdjustment",StratName="Stratification")
}
