% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlsPower.R
\name{print.glsPower}
\alias{print.glsPower}
\title{Print an object of class `glsPower`}
\usage{
\method{print}{glsPower}(x, ...)
}
\arguments{
\item{x}{object of class glsPower}

\item{...}{Arguments to be passed to methods}
}
\value{
Messages, containing information about (at least) power and
significance level
}
\description{
Print an object of class `glsPower`
}
