% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_alloc_Power.R
\name{print.N_alloc_Power}
\alias{print.N_alloc_Power}
\title{Print an object of class \code{N_alloc_Power}}
\usage{
\method{print}{N_alloc_Power}(x, ...)
}
\arguments{
\item{x}{object of class N_alloc_Power}

\item{...}{Arguments to be passed to methods}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an object of class \code{N_alloc_Power}
}
