% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{RandEff_to_icc}
\alias{RandEff_to_icc}
\title{Transform random effects into ICC, CAC, IAC}
\usage{
RandEff_to_icc(sigResid, tau, gamma = 0, psi = 0)
}
\arguments{
\item{sigResid}{Residual standard deviation on individual level}

\item{tau}{standard deviation of random cluster intercept}

\item{gamma}{standard deviation of random time effect}

\item{psi}{standard deviation of random subject specific intercept}
}
\value{
a list containing four named elements (possibly vectors or matrices):
\itemize{
\item \code{icc} intracluster correlation
\item \code{cac} cluster autocorrelation
\item \code{iac} individual autocorrelation
\item \code{sigResid} Residual standard deviation on individual level
}
}
\description{
This function transforms standard deviations of random effects into
intracluster correlation (ICC), cluster autocorrelation (CAC) and
individual autocorrelation (IAC).
It reproduces the formulae in Hooper et al. (2016).
}
\details{
The formulae used are \cr
\deqn{
       \sigma^2_{marg} = \sigma^2_{resid} + \tau^2 + \gamma^2 + \psi^2 \\[1ex]
       ICC = \frac{\tau^2 + \gamma^2}{\sigma^2_{marg}} \\[1ex]
       CAC = \frac{\tau^2}{\tau^2 + \gamma^2} \\[1ex]
       IAC = \frac{\psi^2}{\sigma^2_{resid}+\psi^2}
      }
}
\references{
Hooper, R., Teerenstra, S., de Hoop, E., & Eldridge, S. (2016). \cr
\emph{Sample size calculation for stepped wedge and other longitudinal cluster randomised trials.} \cr
Statistics in medicine, 35(26), 4718-4728. DOI: 10.1002/sim.7028
}
\seealso{
\code{\link[=icc_to_RandEff]{icc_to_RandEff()}}, \code{\link[=RandEff_to_alpha012]{RandEff_to_alpha012()}} or \code{\link[=alpha012_to_RandEff]{alpha012_to_RandEff()}}
}
