% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Functions.R
\name{realpower.simulate}
\alias{realpower.simulate}
\title{realpower.simulate}
\usage{
realpower.simulate(n = 1, nInd = 100, tau, eta = 0, rho = 0, mu0,
  mu1, design, whichModel = c("AGQ", "PQL", "GEE"), nAGQ = 1)
}
\arguments{
\item{n}{helper}
}
\description{
realpower.simulate
}
