% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialWrapper.R
\name{split_sd}
\alias{split_sd}
\title{'split_sd'}
\usage{
split_sd(trtvec, sd)
}
\arguments{
\item{trtvec}{a vector}

\item{sd}{vector of length 2 (or 3).}
}
\value{
a formula for CoxEval()
}
\description{
'split_sd'
}
\examples{
BuildFormula(VarName="Variable",AdjName="additiveAdjustment",StratName="Stratification")
}
