% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{tTestPwr}
\alias{tTestPwr}
\title{Compute Power of a Wald Test}
\usage{
tTestPwr(d, se, df, sig.level = 0.05)
}
\arguments{
\item{d}{numeric, raw effect}

\item{se}{numeric, standard error}

\item{df}{numeric, degrees of freedom of the t-distribution}

\item{sig.level}{numeric, significance level, defaults to 0.05}
}
\value{
a scalar
}
\description{
Computes the power of a scaled Wald test given a standard error,
an effect size, the degrees of freedom of the t-distribution and
a significance level.
Computes the exact power, see second example
}
\examples{
tTestPwr(4,1,10) ; tTestPwr(4,1,30) ; tTestPwr(4,1,Inf)
}
