% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_formulae.R
\name{VarClosed_Kasza}
\alias{VarClosed_Kasza}
\title{Closed formula for treatment variance in open cohort settings}
\usage{
VarClosed_Kasza(trtMat, tau, gamma = 0, psi = 0, sigma, N, chi)
}
\arguments{
\item{trtMat}{a matrix trtMat to define treatment allocation,
where rows and columns correspond to cluster and timepoints, respectively}

\item{tau}{numeric, standard deviation of random intercepts}

\item{gamma}{numeric, random time effect}

\item{psi}{numeric, random subject specific intercept.}

\item{sigma}{numeric, residual error on subject level.}

\item{N}{numeric, number of individuals per cluster.}

\item{chi}{Attrition factor}
}
\value{
numeric, variance of the estimator for treatment effect
}
\description{
From Kasza et al "Sample size and power calculations for open cohort
longitudinal cluster rondomized trials" 2020
}
