% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_formulae.R
\name{VarClosed_Li}
\alias{VarClosed_Li}
\title{Closed formula for treatment variance, with proportional decay}
\usage{
VarClosed_Li(trtMat, tau, psi, N, AR)
}
\arguments{
\item{trtMat}{a matrix trtMat to define treatment allocation,
where rows and columns correspond to cluster and timepoints, respectively}

\item{tau}{numeric, standard deviation of random intercepts}

\item{psi}{numeric, random subject specific intercept.}

\item{N}{numeric, number of individuals per cluster.}

\item{AR}{numeric (scalar), It defines the AR(1)-correlation of random effects.}
}
\value{
numeric, variance of the estimator for treatment effect
}
\description{
From Li et al "Design and analysis considerations for cohort stepped wedge
cluster randomized trials with a decay correlation structure
}
