% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialWrapper.R
\name{wlsGlmmPower}
\alias{wlsGlmmPower}
\title{wlsGlmmPower}
\usage{
wlsGlmmPower(Cl, mu0, mu1, tau, eta = NULL, rho = NULL,
  design = "SWD", timepoints = NULL, family = "binomial", N = NULL,
  sig.level = 0.05, trt_delay = NULL, time_adjust = "factor",
  df_adjust = "none", verbose = TRUE)
}
\arguments{
\item{Cl}{integer (vector), number of clusters per wave (in SWD)}

\item{mu0}{average under control}

\item{mu1}{average under intervention}

\item{tau}{numeric, standard deviation of random intercepts (soon) to be understood on the linear predictor level}

\item{eta}{*not implemented*}

\item{rho}{*not implemented*}

\item{design}{study design. One of the following values: "SWD", "parallel", "parallel_baseline" ...}

\item{family}{currently only "binomial"}

\item{N}{Number of Individuals per cluster/period}

\item{sig.level}{numeric, significance level, defaults to 0.05}

\item{df_adjust}{character, one of the following: **not implemented**}

\item{verbose}{Logical}

\item{Power}{Target power. If between 0 and 1, function returns needed `N`}

\item{delay}{numeric (possibly vector), value between 0 and 1 specifing the
intervention effect in the first (second ... ) intervention phase *not implemented*}
}
\value{
as `wlsMixedPower`
}
\description{
wlsGlmmPower
}
\examples{
wlsGlmmPower(Cl=c(2,2,2),mu0=0.25,mu1=0.5,tau=0.05,N=20)
}
