% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunction.R
\name{wlsMixedPower}
\alias{wlsMixedPower}
\title{'wlsMixedPower'}
\usage{
wlsMixedPower(EffSize, I, sigma, tau, family = gaussian(), N = 1,
  sig.level = 0.05)
}
\arguments{
\item{EffSize}{numeric, raw effect}

\item{I}{integer (vector), number of clusters per wave (in SWD)}

\item{sigma}{numeric, residual error of cluster means if no N given.
Else residual error on individual level}

\item{tau}{numeric, standard deviation of random intercepts}

\item{sig.level}{numeric, significance level, defaults to 0.05}
}
\description{
'wlsMixedPower'
}
