% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlsMixedPower.R
\name{zTestPwr}
\alias{zTestPwr}
\title{zTestPwr}
\usage{
zTestPwr(d, se, sig.level = 0.05)
}
\arguments{
\item{d}{numeric, raw effect}

\item{se}{numeric, standard error}

\item{sig.level}{numeric, significance level, defaults to 0.05}
}
\value{
a scalar
}
\description{
computes the power of a z-test given a standard error, an effect size and a significance level.
Computes the exact power, see second example
}
\examples{
zTestPwr(4,1) :
zTestPwr(qnorm(.975),1) == pnorm(qnorm(.975),qnorm(.975),1) + pnorm(-qnorm(.975),qnorm(.975),1)
}
