% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunction.R
\name{zTestSampSize}
\alias{zTestSampSize}
\title{zTestSampSize}
\usage{
zTestSampSize(d, sd, Power, sig.level = 0.05)
}
\arguments{
\item{d}{numeric, raw effect}

\item{Power}{target power}

\item{sig.level}{numeric, significance level, defaults to 0.05}

\item{se}{numeric, standard deviaton (on individual level)}
}
\value{
a scalar
}
\description{
calculate needed sample size for given target power, effect size and individual variance
does it also work for closed cohorts? i fear not (gotta think about it ...)
}
\examples{
zTestPwr(4,1)
}
