% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_designs.R
\name{compare_designs}
\alias{compare_designs}
\title{compare_designs}
\usage{
compare_designs(
  Cl = NULL,
  mu0,
  mu1,
  sigma,
  tau,
  family = gaussian(),
  timepoints = NULL,
  N = NULL,
  sig.level = 0.05,
  DesMat = NULL,
  trtDelay = NULL
)
}
\arguments{
\item{Cl}{as in wlsMixedPower}

\item{mu0}{as in wlsMixedPower}

\item{mu1}{as in wlsMixedPower}

\item{sigma}{as in wlsMixedPower}

\item{tau}{as in wlsMixedPower}

\item{family}{as in wlsMixedPower}

\item{timepoints}{as in wlsMixedPower}

\item{N}{as in wlsMixedPower}

\item{sig.level}{as in wlsMixedPower}

\item{DesMat}{as in wlsMixedPower}

\item{trtDelay}{as in wlsMixedPower}
}
\description{
a short wrapper that compares parallel, sw, and parallel+ baseline design
}
