% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_designs.R
\name{compare_designs}
\alias{compare_designs}
\title{compare_designs}
\usage{
compare_designs(wlsPower)
}
\arguments{
\item{wlsPower}{object of class `wlsPower`.}
}
\description{
a short wrapper that compares parallel, sw, and parallel+ baseline design
}
