% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_designs.R
\name{compare_designs}
\alias{compare_designs}
\title{compare_designs}
\usage{
compare_designs(EffSize, sigma, tau, family = gaussian(),
  timepoints = NULL, N = NULL, sig.level = 0.05, Cl = NULL,
  delay = NULL, design = "SWD")
}
\description{
for given parameters, calculate power and weights for competing designs. **!top priority!**
 Integrate optimal mix in the sense of Kasza 2017 **low priority**
}
