% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_designs.R
\name{compare_designs2}
\alias{compare_designs2}
\title{Title}
\usage{
compare_designs2(
  Designs,
  mu0,
  mu1,
  family = gaussian(),
  sigma,
  tau,
  eta = NULL,
  rho = NULL,
  sig.level = 0.05
)
}
\arguments{
\item{Designs}{different designs to compare}

\item{mu0}{as in wlsMixedPower}

\item{mu1}{as in wlsMixedPower}

\item{family}{as in wlsMixedPower}

\item{sigma}{as in wlsMixedPower}

\item{tau}{as in wlsMixedPower}

\item{eta}{as in wlsMixedPower}

\item{rho}{as in wlsMixedPower}

\item{sig.level}{as in wlsMixedPower}
}
\value{

}
\description{
Title
}
