% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlsMixedPower.R
\name{compute_wlsPower}
\alias{compute_wlsPower}
\title{compute_wlsPower}
\usage{
compute_wlsPower(
  DesMat,
  EffSize,
  sigma,
  tau,
  N,
  Power,
  CovBlk = NULL,
  df_adjust = df_adjust,
  sig.level,
  verbose
)
}
\arguments{
\item{DesMat}{list, containing a matrix, the design matrix,
numeric timepoints, numeric total number of Clusters}

\item{EffSize}{numeric, raw effect}

\item{sigma}{numeric, residual error of cluster means if no N given.}

\item{tau}{numeric, standard deviation of random intercepts}

\item{N}{integer, number of individuals per cluster.}

\item{Power}{numeric, a specified target power. If supplied, the minimal N is returned.}

\item{df_adjust}{character, one of the following: **not implemented**}

\item{sig.level}{numeric, significance level, defaults to 0.05}

\item{verbose}{logical, should the function return the design and covariance matrix?}
}
\description{
compute_wlsPower
}
