% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlsMixedPower.R
\name{compute_wlsPower}
\alias{compute_wlsPower}
\title{compute_wlsPower}
\usage{
compute_wlsPower(
  DesMat,
  EffSize,
  sigma,
  tau = 0,
  eta = NULL,
  tauAR = NULL,
  etaAR = NULL,
  rho = NULL,
  gamma = NULL,
  psi = NULL,
  N = NULL,
  CovMat = NULL,
  dfAdjust = "none",
  sig.level = 0.05,
  INDIV_LVL = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{DesMat}{list, containing a matrix, the design matrix,
numeric timepoints, numeric total number of Clusters}

\item{EffSize}{numeric, raw effect}

\item{sigma}{numeric, residual error of cluster means if no N given.}

\item{tau}{numeric, standard deviation of random intercepts}

\item{eta}{numeric, standard deviation of random slopes}

\item{tauAR}{numeric (scalar), value between 0 and 1. Defaults to NULL.
If `tauAR` is not NULL, the random intercept
`tau` is AR1-correlated. *Currently not compatible with `rho`!=0 !*}

\item{etaAR}{numeric (scalar), value between 0 and 1. Defaults to NULL.
If `etaAR` is not NULL, the random slope
`eta` is AR1-correlated. *Currently not compatible with `rho`!=0 !*}

\item{rho}{numeric, correlation of tau and eta **not implemented**}

\item{gamma}{numeric (scalar), random time effect}

\item{N}{integer, number of individuals per cluster per timepoint.}

\item{CovMat}{numeric, a positive-semidefinite matrix  with
*#Cluster* \eqn{\cdot} *timepoints* rows/columns. If `CovMat` is given, `sigma`,
`tau`, `eta` and `rho` are ignored.}

\item{dfAdjust}{character, one of the following: **not implemented**}

\item{sig.level}{numeric, significance level, defaults to 0.05}

\item{INDIV_LVL}{logical, Should the computation be done on a individual
(subcluster) level or on cluster means? Defaults to FALSE, since computation
on individual level is quite costly. This option exists mainly for demontration purposes.}

\item{verbose}{logical, should the function return the design and covariance matrix?}
}
\description{
compute_wlsPower
}
