% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_CovMat.R
\name{construct_CovBlk}
\alias{construct_CovBlk}
\title{construct_CovBlk}
\usage{
construct_CovBlk(timepoints, sigma, tau)
}
\arguments{
\item{timepoints}{numeric (scalar), number of timeperiods, Cl.e. dimension of blocks in covariance matrix}

\item{sigma}{numeric (scalar or vector of length `timepoints`),
residual error (usually of cluster means)}

\item{tau}{numeric (scalar or vector of length timepoints), standard deviation of random intercepts,
A vector of length *timepoints* is interpreted as a variing sd over time (also used for binomial outcomes).}
}
\value{
a block of a covariance matrix
}
\description{
constructs the covariance matrix for multiple measurements of one cluster
}
\examples{
construct_CovBlk(timepoints=5, sigma=2, tau=2)
}
