% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_CovMat.R
\name{construct_CovMat}
\alias{construct_CovMat}
\title{Construct a Covariance Matrix}
\usage{
construct_CovMat(
  sumCl = NULL,
  timepoints = NULL,
  sigma,
  tau,
  eta = NULL,
  AR = NULL,
  rho = NULL,
  gamma = NULL,
  trtMat = NULL,
  N = NULL,
  CovBlk = NULL,
  psi = NULL,
  INDIV_LVL = FALSE
)
}
\arguments{
\item{sumCl}{total number of clusters}

\item{timepoints}{numeric (scalar or vector), number of timepoints (periods).
If design is swd, timepoints defaults to length(Cl)+1.
Defaults to 1 for parallel designs.}

\item{sigma}{numeric, residual error of cluster means if no N given.}

\item{tau}{numeric, standard deviation of random intercepts}

\item{eta}{numeric (scalar or matrix), standard deviation of random slopes.
If \code{eta} is given as scalar, \code{trtMat} is needed as well.}

\item{AR}{numeric, vector containing up to three values, each between 0 and 1.
Defaults to NULL. It defines the AR(1)-correlation of random effects.
The first element corresponds to the cluster intercept, the second to the
treatment effect and the third to subject specific intercept.
If only one element is provided, autocorrelation of all random effects is
assumed to be the same.
\emph{Currently not compatible with \code{rho}!=0 !}}

\item{rho}{numeric (scalar), correlation of \code{tau} and \code{eta}. The default is no correlation.}

\item{gamma}{numeric (scalar), random time effect}

\item{trtMat}{a matrix of dimension \emph{#Cluster} x \emph{timepoints} as produced by
the function \code{\link{construct_trtMat}}, indicating the cluster-periods that receive
interventional treatment. Defaults to NULL. If trtMat is given, the arguments
\code{sumCl} and \code{timepoints} are ignored (!).}

\item{N}{numeric, number of individuals per cluster. Either a scalar, vector
of length #Clusters or a matrix of dimension #Clusters x timepoints.
Defaults to 1 if not passed.}

\item{CovBlk}{a matrix of dimension \emph{timepoints} x \emph{timepoints}.}

\item{psi}{numeric (scalar), random subject specific intercept.
Leads to a closed cohort setting}

\item{INDIV_LVL}{logical, should the computation be conducted on an
individual level? This leads to longer run time and is
mainly for diagnostic purposes.}
}
\value{
a covariance matrix
}
\description{
constructs a (block diagonal) covariance matrix.
This function calls \code{\link{construct_CovBlk}}
(or \code{\link{construct_CovSubMat}} in case of repeated
observations of the same individuals) for each block.
}
\examples{

## Two clusters, three timepoints,
## residual standard error sd=3, random slope sd=1.
construct_CovMat(sumCl=2, timepoints=3, sigma=3, tau=1)
##
##
## ... with random slope as AR-1 process
construct_CovMat(sumCl=2, timepoints=3, sigma=3, tau=1, AR=.8)
##
##
## ... with sigma and tau variing over time and between clusters:
construct_CovMat(sumCl=2,timepoints=3,
                 sigma=matrix(c(1,2,2,1,1,2),nrow=2, byrow=TRUE),
                 tau=matrix(c(.2,.1,.1,.2,.2,.1),nrow=2, byrow=TRUE),
                 N=c(3,4))
}
