% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_CovMat.R
\name{construct_CovMat}
\alias{construct_CovMat}
\title{'construct_CovMat'}
\usage{
construct_CovMat(SumCl, timepoints = NULL, sigma, tau,
  family = gaussian(), N = NULL)
}
\arguments{
\item{SumCl}{total number of clusters}

\item{timepoints}{numeric, scalar}

\item{sigma}{numeric, residual error of cluster means if no N given.
Else residual error on individual level}

\item{tau}{numeric, standard deviation of random intercepts}

\item{family}{distribution , not implemented}

\item{N}{integer (vector), number of individuals per cluster.
Defaults to 'rep(1,sum(I))' if not passed.}
}
\value{
a formula for CoxEval()
}
\description{
'construct_CovMat'
}
\examples{
BuildFormula(VarName="Variable",AdjName="additiveAdjustment",StratName="Stratification")
}
