% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_CovMat.R
\name{construct_CovMat}
\alias{construct_CovMat}
\title{construct_CovMat}
\usage{
construct_CovMat(SumCl, timepoints, sigma, tau, N = NULL)
}
\arguments{
\item{SumCl}{total number of clusters}

\item{timepoints}{numeric, scalar}

\item{sigma}{numeric, residual error of cluster means if no N given.
Else residual error on individual level}

\item{tau}{numeric, standard deviation of random intercepts}

\item{N}{integer (vector), number of individuals per cluster.
Defaults to 'rep(1,sum(Cl))' if not passed.}
}
\value{
a covariance matrix
}
\description{
constructs a covariance matrix, a bloc diagonal matrix. Calls `construct_CovBlk`
for each block.
}
\examples{
construct_CovMat(SumCl=2,timepoints=3, sigma=list(c(1,2,2),c(1,1,2)),tau=list(c(.2,.1,.1),c(.2,.2,.1)),N=c(20,16))
}
