% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_CovMat.R
\name{construct_CovMat}
\alias{construct_CovMat}
\title{construct_CovMat}
\usage{
construct_CovMat(
  SumCl = NULL,
  timepoints = NULL,
  sigma,
  tau,
  eta = NULL,
  rho = NULL,
  trtMat = NULL,
  N = NULL,
  CovBlk = NULL
)
}
\arguments{
\item{SumCl}{total number of clusters}

\item{timepoints}{numeric, scalar}

\item{sigma}{numeric (scalar, vector or matrix), residual error of cluster means.}

\item{tau}{numeric (scalar, vector or matrix), standard deviation of random intercepts}

\item{eta}{numeric (scalar), standard deviation of random slopes. If `eta` is
given, `trtMat` is needed as well.}

\item{rho}{numeric (scalar), correlation of random effects `tau` and `eta`. **not implemented**}

\item{trtMat}{a matrix of dimension *#Cluster* x *timepoints* as produced by
the function `construct_trtMat`, indicating the cluster-periods that receive
interventional treatment. Defaults to NULL. If trtMat is given, the arguments
`SumCl` and `timepoints` are ignored (!).}

\item{N}{numeric (scalar, vector or matrix), number of individuals per cluster.
Defaults to 'rep(1,sum(Cl))' if not passed.}

\item{CovBlk}{a matrix of dimension *timepoints* x *timepoints*.}
}
\value{
a covariance matrix
}
\description{
constructs a covariance matrix, a bloc diagonal matrix. Calls `construct_CovBlk`
for each block.
}
\examples{
construct_CovMat(SumCl=2,timepoints=3,
                 sigma=list(c(1,2,2),c(1,1,2)),
                 tau=list(c(.2,.1,.1),c(.2,.2,.1)),N=c(20,16))
}
