% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_CovMat.R
\name{construct_CovSubMat}
\alias{construct_CovSubMat}
\title{construct_CovSubMat}
\usage{
construct_CovSubMat(
  N,
  timepoints,
  sigma,
  tau,
  eta = NULL,
  tauAR = NULL,
  etaAR = NULL,
  rho = NULL,
  gamma = 0,
  trtMat = NULL,
  psi = NULL,
  INDIV_LVL = FALSE
)
}
\arguments{
\item{N}{Number of individuals per cluster}

\item{timepoints}{Number of timepoints}

\item{sigma}{numeric (vector of length `timepoints`),
residual error (usually of cluster means)}

\item{tau}{numeric (vector of length *timepoints*), standard deviation of random intercepts,
A vector of length *timepoints* is interpreted as a variing sd over time (also used for binomial outcomes).}

\item{eta}{numeric (vector of length *timepoints*), standard deviation of random slope}

\item{tauAR}{numeric (scalar), value between 0 and 1. Defaults to NULL. If `tauAR` is not NULL, the random intercept
`tau` is AR1-correlated. *Currently not compatible with `rho`!=0 !*}

\item{etaAR}{numeric (scalar), value between 0 and 1. Defaults to NULL. If `etaAR` is not NULL, the random slope
`eta` is AR1-correlated. *Currently not compatible with `rho`!=0 !*}

\item{rho}{numeric (scalar), correlation of random effects `tau` and `eta`.}

\item{gamma}{numeric (scalar), standard deviation of a random time effect.}

\item{trtMat}{a matrix of dimension *#Cluster* x *timepoints* as produced by
the function `construct_trtMat`, indicating the cluster-periods that receive
interventional treatment. Defaults to NULL. If trtMat is given, the arguments
`SumCl` and `timepoints` are ignored (!).}

\item{psi}{random individuum effect. Leads to a closed cohort setting}
}
\value{

}
\description{
construct_CovSubMat
}
