% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_DesMat.R
\name{construct_DesMat}
\alias{construct_DesMat}
\title{construct_DesMat}
\usage{
construct_DesMat(Cl, trt_delay = NULL, design = "SWD",
  timepoints = NULL, time_adjust = "factor")
}
\arguments{
\item{Cl}{integer (vector), number of clusters per wave (in SWD)}

\item{trt_delay}{*implemented for swd, parallel and parallel w baseline*}

\item{timepoints}{numeric, scalar}

\item{time_adjust}{character, specifies adjustment for time periods. Defaults to "factor".}
}
\value{
a matrix (for a stepped wedge design)
}
\description{
constructs the design matrix.
}
\details{
Note: Unlike the usual notiation, the treatment is in the first column (for easier access by higher level functions).
}
\examples{
construct_DesMat(Cl=c(2,0,1))

}
