% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_DesMat.R
\name{construct_DesMat}
\alias{construct_DesMat}
\title{construct_DesMat}
\usage{
construct_DesMat(Cl, delay = NULL, design = "SWD")
}
\arguments{
\item{Cl}{integer (vector), number of clusters per wave (in SWD)}

\item{delay}{*not implemented*}
}
\value{
a design matrix (for a stepped wedge design)
}
\description{
constructs a
}
\examples{
construct_DesMat(Cl=c(2,0,1))

}
