% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_DesMat.R
\name{construct_DesMat}
\alias{construct_DesMat}
\title{construct_DesMat}
\usage{
construct_DesMat(Cl, delay = NULL, design = "SWD",
  timepoints = timepoints)
}
\arguments{
\item{Cl}{integer (vector), number of clusters per wave (in SWD)}

\item{delay}{*not implemented*}

\item{timepoints}{numeric, scalar}
}
\value{
a matrix (for a stepped wedge design)
}
\description{
constructs the design matrix.
}
\examples{
construct_DesMat(Cl=c(2,0,1))

Note: Unlike the usual notiation, the treatment is in the first column (for easier access by higher level functions).
}
