% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_DesMat.R
\name{construct_DesMat}
\alias{construct_DesMat}
\title{construct_DesMat}
\usage{
construct_DesMat(
  Cl = NULL,
  trtDelay = NULL,
  design = "SWD",
  timepoints = NULL,
  timeAdjust = "factor",
  period = NULL,
  trtmatrix = NULL,
  timeBlk = NULL
)
}
\arguments{
\item{Cl}{integer (vector), number of clusters per wave (in SWD)}

\item{trtDelay}{numeric (possibly vector), value(s) between 0 and 1 specifing the
intervention effect in the first (second ... ) intervention phase}

\item{design}{character, specifies the study design. Defaults to "SWD".}

\item{timepoints}{numeric, scalar}

\item{timeAdjust}{character, specifies adjustment for time periods. Defaults to "factor".}

\item{period}{}
}
\value{
a matrix (for a stepped wedge design)
}
\description{
constructs the design matrix.
}
\details{
Note: Unlike the usual notiation, the treatment is in the first column (for easier access by higher level functions).
}
\examples{
construct_DesMat(Cl=c(2,0,1))

}
