% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_DesMat.R
\name{construct_DesMat}
\alias{construct_DesMat}
\title{'construct_DesMat'}
\usage{
construct_DesMat(I, delay = NULL)
}
\arguments{
\item{I}{integer (vector), number of clusters per wave (in SWD)}

\item{delay}{*not implemented*}
}
\value{
a formula for CoxEval()
}
\description{
'construct_DesMat'
}
\examples{
BuildFormula(VarName="Variable",AdjName="additiveAdjustment",StratName="Stratification")
}
