% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_DesMat.R
\name{construct_timeadjust}
\alias{construct_timeadjust}
\title{construct_timeadjust}
\usage{
construct_timeadjust(Cl, timepoints, time_adjust)
}
\arguments{
\item{Cl}{integer (vector), number of clusters per wave (in SWD)}

\item{timepoints}{numeric, scalar}

\item{time_adjust}{character, specifies adjustment for time periods. Defaults to "factor".}
}
\value{
What is returned? TODO
}
\description{
construct_timeadjust
}
\examples{

}
