% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_DesMat.R
\name{construct_trtMat}
\alias{construct_trtMat}
\title{construct_trtMat}
\usage{
construct_trtMat(Cl, trtDelay, dsntype, timepoints = NULL)
}
\arguments{
\item{Cl}{integer (vector), number of clusters per wave (in SWD)}

\item{trtDelay}{numeric (possibly vector), value(s) between 0 and 1 specifying the
intervention effect in the first (second ... ) intervention phase}

\item{dsntype}{character, specifies the study design. Defaults to "SWD".}

\item{timepoints}{numeric, scalar}
}
\value{
a matrix trtMat
}
\description{
construct_trtMat
}
\examples{
construct_trtMat(Cl=c(1,2,1), trtDelay=c(.2,.8), dsntype="SWD")


}
