% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunction.R
\name{optFunction}
\alias{optFunction}
\title{optFunction}
\usage{
optFunction(DesMat, EffSize, sigma, tau, N, Power, df_adjust, sig.level)
}
\arguments{
\item{DesMat}{list, containing a matrix, the design matrix,
numeric timepoints, numeric total number of Clusters}

\item{EffSize}{numeric, raw effect}

\item{sigma}{numeric, residual error of cluster means if no N given.}

\item{tau}{numeric, standard deviation of random intercepts}

\item{N}{integer, number of individuals per cluster.}

\item{Power}{numeric, a specified target power. If supplied, the minimal N is returned.}

\item{sig.level}{numeric, significance level, defaults to 0.05}
}
\description{
only to be called by wlsMixedPower.
}
