% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_DesMat.R
\name{plot.DesMat}
\alias{plot.DesMat}
\title{plot.DesMat}
\usage{
\method{plot}{DesMat}(x, ...)
}
\arguments{
\item{x}{d}

\item{...}{Arguments to be passed to methods}
}
\value{
a plotly html widget, displaying the treatment status
}
\description{
plot.DesMat
}
