% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_DesMat.R
\name{plot.DesMat}
\alias{plot.DesMat}
\title{plot.DesMat}
\usage{
\method{plot}{DesMat}(x, show_colorbar = FALSE, INDIV_LVL = FALSE, ...)
}
\arguments{
\item{x}{An object of class `DesMat`}

\item{show_colorbar}{logical, should the colorbar be shown?}

\item{INDIV_LVL}{logical, should the computation be conducted on an
individual level? This leads to longer run time and is
mainly for diagnostic purposes.}

\item{...}{Arguments to be passed to methods}
}
\value{
a plotly html widget, displaying the treatment status
}
\description{
plot.DesMat
}
\examples{
x <- construct_DesMat(Cl=c(2,2,2,0,2,2,2),.5)
}
