% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlsPower.R
\name{plot.wlsPower}
\alias{plot.wlsPower}
\title{plot an object of class `wlsPower`}
\usage{
\method{plot}{wlsPower}(
  x,
  which = NULL,
  show_colorbar = NULL,
  annotations = NULL,
  marginal_plots = TRUE,
  ...
)
}
\arguments{
\item{x}{object of class wlsPower}

\item{which}{Specify a subset of the numbers `1:4` to select plots. The default is
`1:2` or `1`, depending on whether `x` contains the information content.}

\item{show_colorbar}{logical, should the colorbars be shown?}

\item{annotations}{logical, should the cell contributions be annotated in the Plot?}

\item{marginal_plots}{should the influence of whole periods, clusters also be plotted?}

\item{...}{Arguments to be passed to methods}
}
\value{
a list of plotly html widgets
}
\description{
Up to four plots (selectable by `which`) that visualise:
the contribution of each cluster-period cell to the treatment effect estimator,
the information content of each cluster-period cell,
the treatment status for each cluster for each time point and
the covariance matrix. By default, only the first two plots are returned.
}
