% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlsMixedPower.R
\name{plot.wlsPower}
\alias{plot.wlsPower}
\title{plot.wlsPower}
\usage{
\method{plot}{wlsPower}(x, ...)
}
\arguments{
\item{x}{w}

\item{...}{Arguments to be passed to methods}
}
\description{
plot.wlsPower
}
