% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wlsPower.R
\name{plot.wlsPower}
\alias{plot.wlsPower}
\title{plot.wlsPower}
\usage{
\method{plot}{wlsPower}(wlsPower)
}
\arguments{
\item{wlsPower}{Output of function wlsMixedPower}
}
\value{
three ggplot2 objects.
}
\description{
visual interpretation of cluster and period influences on intervention estimate
}
\examples{
# wlsPowerOut <- wlsMixedPower(EffSize=.1,Cl=c(1,2,1,0,1),sigma=.2,tau=0.05,N=10,verbose=TRUE)
# plot_wlsPower(wlsPowerOut)

}
