% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlsPower.R
\name{plot_InfoContent}
\alias{plot_InfoContent}
\title{plot the information content of a wls object}
\usage{
plot_InfoContent(IC)
}
\arguments{
\item{IC}{a matrix with information content for each cluster at each time period}
}
\value{
a plotly object
}
\description{
plot the information content of a wls object
}
