% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wlsPower.R
\name{plot_wlsPower}
\alias{plot_wlsPower}
\title{'plot_wlsPower'}
\usage{
plot_wlsPower(wlsPower)
}
\arguments{
\item{wlsPower}{Output of function wlsMixedPower}
}
\value{
three ggplot2 objects.
}
\description{
visual interpretation of cluster and period influences on intervention estimate
}
\examples{
wlsPowerOut <- wlsMixedPower(EffSize=.1,I=c(1,2,1,0,1),sigma=.2,tau=0.05,N=10)
plot_wlsPower(wlsPowerOut)
}
