% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_DesMat.R
\name{print.DesMat}
\alias{print.DesMat}
\title{print.DesMat}
\usage{
\method{print}{DesMat}(x, ...)
}
\arguments{
\item{x}{An object of class `DesMat}

\item{...}{Arguments to be passed to methods}
}
\value{
Messages with information about the design.
}
\description{
print.DesMat
}
