% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlsPower.R
\name{print.wlsPower}
\alias{print.wlsPower}
\title{Print an object of class `wlsPower`}
\usage{
\method{print}{wlsPower}(x, ...)
}
\arguments{
\item{x}{object of class wlsPower}

\item{...}{Arguments to be passed to methods}
}
\value{
Messages, containing information about (at least) power and
significance level
}
\description{
Print an object of class `wlsPower`
}
