% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlsMixedPower.R
\name{print.wlsPower}
\alias{print.wlsPower}
\title{print.wlsPower}
\usage{
\method{print}{wlsPower}(x, ...)
}
\arguments{
\item{x}{w}

\item{...}{Arguments to be passed to methods}
}
\description{
print.wlsPower
}
