% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Functions.R
\name{Simslim}
\alias{Simslim}
\title{SimSlim}
\usage{
Simslim(design, n, mu0, mu1, time.effect, sigma, tau, eta, rho = 0,
  time.lab = NULL, seed = NULL, seed.designonly = F)
}
\arguments{
\item{seed.designonly}{}
}
\description{
SimSlim
}
