% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialWrapper.R
\name{split_sd}
\alias{split_sd}
\title{'split_sd'}
\usage{
split_sd(trtvec, sd)
}
\arguments{
\item{trtvec}{a vector}

\item{sd}{vector of length 2 (or 3).}
}
\value{
numeric, scalar.
}
\description{
small helper function for wlsGlmmPower. returns sd_0 for periods in control and sd_1
 for interventional periods for each cluster.
}
