% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SteppedPower-pkg.R
\docType{package}
\name{SteppedPower-pkg}
\alias{SteppedPower-pkg}
\title{SteppedPower}
\description{
SteppedPower offers tools for power and sample size
calculation as well as design diagnostics for
longitudinal mixed model settings, with a focus on stepped wedge designs.
All calculations are oracle estimates i.e. assume random effect variances
to be known (or guessed) in advance.
}
\author{
Philipp Mildenberger \email{pmildenb@uni-mainz.de}
}
