% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SteppedPower-pkg.R
\docType{package}
\name{SteppedPower-pkg}
\alias{SteppedPower-pkg}
\title{SteppedPower}
\description{
SteppedPower offers options for power calculation as well
as sample size calculation (by the use of uniroot() ) for
longitudinal mixed model settings with up to two levels of clustering.
All calculations are oracle estimates i.e. assume random effect variances
to be known (or guessed) in advance.
}
\author{
Philipp Mildenberger \email{pmildenb@uni-mainz.de}
}
