% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SteppedPower-pkg.R
\docType{package}
\name{SteppedPower-pkg}
\alias{SteppedPower-pkg}
\alias{SteppedPower-pkg-package}
\title{SteppedPower}
\description{
Assesses Power of mixed models for the evaluation of a intervention.
}
\author{
Philipp Mildenberger \email{pmildenb@uni-mainz.de}
}
