% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SteppedPower-pkg.R
\docType{package}
\name{SteppedPower-pkg}
\alias{SteppedPower-pkg}
\alias{SteppedPower-pkg-package}
\title{SteppedPower}
\description{
what it does needs further explanation, urgently
It's but a one-trick pony.
Neuer sinnarmer Satz.
}
\author{
Philipp Mildenberger \email{pmildenb@uni-mainz.de}
}
