% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SteppedPower-pkg.R
\docType{package}
\name{SteppedPower-pkg}
\alias{SteppedPower-pkg}
\title{SteppedPower}
\description{
The package `SteppedPower` aims at providing tools for power - and sample
size - calculation and design diagnostics in interventional study designs,
with a special focus on stepped wedge designs.
All calculations are oracle estimates i.e. assume random effect variances
to be known (or guessed) in advance.
}
\author{
Philipp Mildenberger \email{pmildenb@uni-mainz.de}
}
