% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlsGlmmPower.R
\name{wlsGlmmPower}
\alias{wlsGlmmPower}
\title{wlsGlmmPower}
\usage{
wlsGlmmPower(
  Cl,
  mu0,
  mu1,
  marginal_mu = TRUE,
  tau = NULL,
  tau_lin = NULL,
  eta = NULL,
  rho = NULL,
  design = "SWD",
  timepoints = NULL,
  family = "binomial",
  N = NULL,
  sig.level = 0.05,
  trtDelay = NULL,
  timeAdjust = "factor",
  df_adjust = "none",
  verbose = TRUE
)
}
\arguments{
\item{Cl}{integer (vector), number of clusters per wave (in SWD)}

\item{mu0}{numeric, (marginal) average under control}

\item{mu1}{numeric, (marginal) average under intervention}

\item{tau}{numeric, standard deviation of random intercepts, interpreted as normal noise on the proportions}

\item{tau_lin}{numeric, standard deviation of random intercepts on the linear predictor}

\item{eta}{*not implemented*}

\item{rho}{*not implemented*}

\item{design}{study design. One of the following values: "SWD", "parallel", "parallel_baseline" ...}

\item{family}{currently only "binomial"}

\item{N}{Number of Individuals per cluster/period}

\item{sig.level}{numeric, significance level, defaults to 0.05}

\item{df_adjust}{character, one of the following: **not implemented**}

\item{verbose}{logical, shall the function be chatty about what it did?}

\item{Power}{Target power. If between 0 and 1, function returns needed `N`}

\item{delay}{numeric (possibly vector), value between 0 and 1 specifing the
dose of intervention in the first (second ... ) intervention phase *not implemented*}
}
\value{
as `wlsMixedPower`
}
\description{
wlsGlmmPower
}
