% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialWrapper.R
\name{wlsGlmmPower}
\alias{wlsGlmmPower}
\title{'wlsGlmmPower'}
\usage{
wlsGlmmPower(I, mu0, mu1, tau, eta = NULL, rho = NULL,
  family = "binomial", N = NULL, sig.level = 0.05, delay = NULL)
}
\arguments{
\item{I}{integer (vector), number of clusters per wave (in SWD)}

\item{mu0}{average under control}

\item{mu1}{average under intervention}

\item{tau}{numeric, standard deviation of random intercepts (soon) to be understood on the linear predictor level}

\item{sig.level}{numeric, significance level, defaults to 0.05}

\item{delay}{numeric (possibly vector), value between 0 and 1 specifing the
intervention effect in the first (second ... ) intervention phase *not implemented*}
}
\value{

}
\description{
'wlsGlmmPower'
}
\examples{
wlsGlmmPower(I=c(2,2,2),mu0=0.25,mu1=0.5,tau=0.05,N=20)
}
