% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunction.R
\name{wlsMixedPower}
\alias{wlsMixedPower}
\title{'wlsMixedPower'}
\usage{
wlsMixedPower(EffSize, sigma, tau, family = gaussian(),
  timepoints = NULL, N = NULL, sig.level = 0.05, DesMat = NULL,
  I = NULL, delay = NULL)
}
\arguments{
\item{EffSize}{numeric, raw effect}

\item{sigma}{numeric, residual error of cluster means if no N given.
Else residual error on individual level}

\item{tau}{numeric, standard deviation of random intercepts}

\item{sig.level}{numeric, significance level, defaults to 0.05}

\item{I}{integer (vector), number of clusters per wave (in SWD)}

\item{delay}{numeric (possibly vector), value between 0 and 1 specifing the
intervention effect in the first (second ... ) intervention phase  *not implemented*}
}
\value{
a formula for CoxEval()
}
\description{
'wlsMixedPower'
}
\examples{
BuildFormula(VarName="Variable",AdjName="additiveAdjustment",StratName="Stratification")
}
