% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlsMixedPower.R
\name{wlsMixedPower}
\alias{wlsMixedPower}
\title{wlsMixedPower}
\usage{
wlsMixedPower(Cl = NULL, timepoints = NULL, DesMat = NULL,
  trt_delay = NULL, time_adjust = "factor", design = "SWD", EffSize,
  sigma, tau, eta = NULL, rho = NULL, CovBlk = NULL, N = NULL,
  Power = NULL, N_range = c(1, 1000), sig.level = 0.05,
  df_adjust = "none", verbose = FALSE)
}
\arguments{
\item{Cl}{integer (vector), number of clusters per wave (in SWD),
or number in control and intervention (in parallel designs)
Else residual error on individual level}

\item{timepoints}{numeric (scalar or vector), number of timepoints (periods).
If design is swd, timepoints defaults to length(Cl)+1. Defaults to 1 for parallel designs.}

\item{DesMat}{matrix of dimension ... , if supplied, timepoints and Cl are ignored.}

\item{trt_delay}{numeric (possibly vector), value(s) between 0 and 1 specifing the
intervention effect in the first (second ... ) intervention phase}

\item{time_adjust}{character, specifies adjustment for time periods. Defaults to "factor".}

\item{design}{character, defines the type of design. Options are "SWD" and "parallel", defaults to "SWD".}

\item{EffSize}{numeric, raw effect}

\item{sigma}{numeric, residual error of cluster means if no N given.}

\item{tau}{numeric, standard deviation of random intercepts}

\item{eta}{numeric, standard deviation of random slopes **not implemented**}

\item{rho}{numeric, correlation of tau and eta **not implemented**}

\item{N}{integer, number of individuals per cluster.}

\item{Power}{numeric, a specified target power. If supplied, the minimal N is returned.}

\item{N_range}{numeric, vector specifiing the lower and upper bound for N, ignored if Power is NULL.}

\item{sig.level}{numeric, significance level, defaults to 0.05}

\item{df_adjust}{character, one of the following: **not implemented**}

\item{verbose}{logical, should the function return the design and covariance matrix?}
}
\value{
a list. First element is the power,
second element the weights of cluster per period for estimating the treatment effect **needs update**
}
\description{
This is the work-horse function of the SteppedPower package.
It calls the constructor functions for the design matrix $X$ and covariance matrix $V$,
and then calculates the variance of the intervention effect via
$$Var(trteff)=(X'V^{-1}X)^{-1}[1,1]$$
}
\examples{
wlsMixedPower(EffSize=1,Cl=c(1,1,1,1,1),sigma=2 ,        tau=0.2, N=c(1,1,1,1,1) )
wlsMixedPower(EffSize=1,Cl=c(1,1,1,1,1),sigma=2*sqrt(2) ,tau=0.2, N=c(2,2,2,2,2) )
}
