% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunction.R
\name{wlsMixedPower}
\alias{wlsMixedPower}
\title{wlsMixedPower}
\usage{
wlsMixedPower(EffSize, sigma, tau, family = gaussian(),
  timepoints = NULL, N = NULL, sig.level = 0.05, DesMat = NULL,
  Cl = NULL, delay = NULL, design = "SWD")
}
\arguments{
\item{EffSize}{numeric, raw effect}

\item{sigma}{numeric, residual error of cluster means if no N given.
Else residual error on individual level}

\item{tau}{numeric, standard deviation of random intercepts}

\item{sig.level}{numeric, significance level, defaults to 0.05}

\item{Cl}{integer (vector), number of clusters per wave (in SWD),
or number in control and intervention (in parallel designs)}

\item{delay}{numeric (possibly vector), value between 0 and 1 specifing the
intervention effect in the first (second ... ) intervention phase  *not implemented*}

\item{design}{character, defines the type of design. Options are "SWD" and "parallel", defaults to "SWD".}
}
\value{
a list of lenght two. First element is the power,
second element the weights of cluster per period for estimating the treatment effect
}
\description{
This is the work-horse function of the SteppedPower package.
It calls the constructor functions for the design matrix $X$ and covariance matrix $V$,
and then calculates the variance of the intervention effect via
$$Var(\theta)=(X'V^{-1}X)^{-1}[1,1]$$
}
\examples{
wlsMixedPower(EffSize=1,Cl=c(1,1,1,1,1),sigma=2 ,        tau=0.2, N=c(1,1,1,1,1) )
wlsMixedPower(EffSize=1,Cl=c(1,1,1,1,1),sigma=2*sqrt(2) ,tau=0.2, N=c(2,2,2,2,2) )
}
