% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunction.R
\name{zTestPwr}
\alias{zTestPwr}
\title{'zTestPower'}
\usage{
zTestPwr(d, se, sig.level = 0.05)
}
\arguments{
\item{d}{numeric, raw effect}

\item{se}{numeric, standard error}

\item{sig.level}{numeric, significance level, defaults to 0.05}
}
\description{
'zTestPower'
}
