% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{zTestSampSize}
\alias{zTestSampSize}
\title{zTestSampSize}
\usage{
zTestSampSize(d, sd, Power, sig.level = 0.05)
}
\arguments{
\item{d}{numeric, raw effect}

\item{sd}{numeric, standard deviaton (on individual level)}

\item{Power}{target power}

\item{sig.level}{numeric, significance level, defaults to 0.05}
}
\value{
a scalar
}
\description{
calculate needed sample size for given target power, effect size and individual variance
does it work ? i fear not (got to think about it ...)
}
\examples{
zTestPwr(4,1)
}
